use RobotoMono font for error images
Follow up to #967. Use RobotoMono font for the png error
image utility as well as the graphs.
